Refactor: world package
